joule_typeck

Struct TypeChecker

pub struct TypeChecker { /* private fields */ }

Type checker state

Implementations

impl TypeChecker

pub fn new() -> Self

pub fn mono_collector(&self) -> &MonoCollector

Get the monomorphization collector

pub fn mono_collector_mut(&mut self) -> &mut MonoCollector

Get mutable access to the monomorphization collector

pub fn set_energy_budget_checking(&mut self, enabled: bool)

Enable or disable energy budget checking

pub fn set_file_to_module(&mut self, map: HashMap<u32, Symbol>)

Set the file-to-module mapping for cross-module type resolution

pub fn type_to_module(&self) -> &HashMap<Symbol, Symbol>

Get the type-to-module mapping for C emission with qualified names

pub fn conflicting_types(&self) -> &HashSet<Symbol>

Get the set of conflicting type names (need module qualification)

pub fn energy_estimator(&self) -> &EnergyEstimator

Get a reference to the energy estimator

pub fn energy_estimator_mut(&mut self) -> &mut EnergyEstimator

Get a mutable reference to the energy estimator

pub fn register_stdlib_module(&mut self, module: &str) -> usize

Register all exports from a stdlib module into the current scope.

Called by the compiler driver when use std::module::* is encountered. Returns the number of items registered (0 if module is unknown or virtual).

pub fn check_file(&mut self, file: &File) -> Result<Hir, Vec<Diagnostic>>

Type check a complete file

pub fn register_type(&mut self, name: Symbol, def_id: HirId)

Register a type in the current scope

pub fn find_impls_for_type( &self, trait_name: Symbol, ty: &Ty, ) -> Vec<&ImplEntry>

Find all implementations of a trait for a given type